:bug: Kill all shell processes
There would be remaining shell processes when Atom closes. This is
because TerminalPlusPanel would call destroy() on every
TerminalPlusView which took too much time. Now the panel destroy method
only kills the shell process and terminal view because the entire
TPView destroy method is too much for a quick shutdown.