:bug: Fix contentEditable bug
On right click, the terminal div will enable contentEditting. This plus
padding or margin on the terminal styling will cause the contents to be
double spaced. The contentEditable is only enabled for a millisecond so
the terminal appears to flash for a moment and then return to normal.
Removed margin and padding in favor of text-align: center to fix.