:art: Refactor status-bar sorting
Replaced jQuery’s sortable interface with the HTML5 drag and drop
interface for the status bar. This is less intensive and removes a
dependency from the project. As a result, status icons have their own
class now and several status bar methods have been reworked for the new
system. Removed the option to disable the sortable interface.