Use deferred objects in toggle
Toggle, open and hide now rely on deferred objects to prevent the
calling of toggle in the middle of an animation. Holding the toggle
keycap would cause the terminal open/hide to get stuck. These methods
now rely on a deferred to signify when something is currently
animating. Destroy now hides the terminal view and then calls destroy
once the hide animation is done.