Add maximize button and return last focus
Add a maximize button for when you quickly need the full size of the
screen to look at something. The button changes to a minimize button at
that point and the terminal view is no longer resizable. Focus now
returns to the last object that had it when the terminal closes. This
allows you to continue editing where you left off.