Refactor animations
Move from deferred objects to listening to webkitTransitionEnd. This is
a lot easier than dealing with resolved and unresolved objects. It is
easier to have methods launch after the transition. Now the
post-animation methods can be in open and hide rather than in toggle.
Change style attribute clearing to style height clearing.