Replace max height with default height
No reason to enforce a maximum height anymore with resizing and the
maximize button. Replace maximumPanelHeight with defaultPanelHeight.
Calculate the maximum height with the sum of the previous height and
the height of the available space left over. Adjust height of xterm now
in order to prevent issues with resizing the whole bottom panel. This
would sometimes not allocate enough space for the toolbar. Adjust
maximize button switch to reflect new button structure. Disable
transition on resize.