:bug: Fix shell launch bugs
Pass the full shell path to login in case the shell isn’t in the
standard PATH variable. Pass in a filteredEnv when creating a terminal
incase the login command could not be run. Remove initial columns and
rows, we will resize on open. Try to fork with given shell, if it fails
use the default shell. Remove on close, we aren’t doing anything with
it anyway.