Add insert-selected-text command
Add insert-selected-text command for inserting text from the text editor
into the active terminal. Default command is ctrl-enter. If there is
selected text, then insert it as a run command. If there is no selected
text, insert the whole line where the cursor is and proceed to the next line.