Detect system language on OS X
Use plutil to read the GlobalPreferences.plist on Macs. We can read in
the file as JSON and parse it to extract the AppleLocale key. Attach
UTF-8 to make it unicode. If the process fails at any point then the
error will be caught, leaving the default of en_US.UTF-8 in place.