Add auto switch terminal (terminal mapping)
Users can now set Terminal-Plus to map to files or folders and switch
terminal views bases on what file they are looking at. If set to file,
the active terminal will swap with the terminal for the file. If set to
folder, the active terminal will swap with a terminal that matches the
folder your file is in. There is now an auto open option for the
mapping.