:bug: Fix Atom relative path issue
In certain instances, Atom will set the project directory to
atom://config, an Atom relative path. That causes issues when setting
the present working directory since the system does not recognize these
paths. On Linux and OS X, the pwd is set to root. On Windows, it just
throws an error.