Add support for IME and special characters
Currently the terminal does not support dead keys or IME input. It would
be too difficult to implement this through Javascript. Users can now
input these characters and use IME via the input dialog. The dialog can
be accessed via the button or terminal-plus:input-dialog command. The
input box in the dialog allows for IME and dead keys. The dialog will
input the text on 'enter' or dismiss itself on 'esc'.