Map terminals while terminals are hidden
Improve the terminal mapping mechanism by setting the pwd and on
terminal initialize. This allows for terminals to be mapped on creation,
rather than on display. Terminals are now switched and created in the
background. The proper terminal is now displayed on toggle, regardless
of whether there is an open terminal panel or not.