:art: Refactor status-bar
Refactor status-bar code into a core module and status-bar view. The
core now handles the main organization and management of the terminals.
The status-bar acts as a view for displaying the status-icons. The
only functionality left in the status-bar is the handling of
status-icon organization and status-icon drop on tab-bar view.