Refactors toggle and removes console log
Move toggle to the terminal-view class because the method is the same
for the panel-view and tab-view. As a result, the tab-view attach method
had to be adjusted. In addition there were minor changes to the
panel-view and tab-view initialize methods. An extra log statement has
been removed as well.