Fix deprecation warning thrown in Atom Beta 1.13 (#129)
`item.getPath()` returns undefined, causing `path.dirname` to throw a warning.  This happens when you create a new file, since we know this is a new file (with no path, unsaved) we can also remove the `item.getTitle() isnt 'untitled'` check as we'll have exited above in the guard clause.